==> ms_notebooks/3-MergedRuns_Fasttree.ipynb <==
# Fast Tree of DADA2 ASVs
* Tree building from alignment using DADA2 sequences from three hyphosphere experiments.  
* Fasta has already been thresholded to reduce computational demand
* Set kernel to Python 2

workDir = '/home/be68/Hyphosphere/data/3Exp/Fasttree/'
OTUFileDir = '/home/be68/Hyphosphere/data/3Exp/'
OTUFile = '/home/be68/Hyphosphere/data/3Exp/phyloseq/seqs_thresh.fasta'


nprocs = 20

import os
import numpy as np
#import entrez.direct
from cogent.app.fasttree import build_tree_from_alignment
from cogent import DNA, LoadSeqs
from Bio import Entrez, Phylo
Entrez.email = "be68@cornell.edu"
#from IPython.display import display, Image, SVG
#from cogent3 import LoadSeqs, FastTree


if not os.path.isdir(workDir):
    os.mkdir(workDir)

!cd $workDir

!pwd

!cd $workDir; ln -f -s $OTUFile

!cd $workDir; head $OTUFile

!printf "Number of OTUs in fasta: "
!cd $workDir; grep -c ">" $OTUFile

## Using SSU-Align to align seqs and masking based on alignment posterior probabilities.

!cd $workDir; ssu-prep -f -x -b 50 --rfonly --dna $OTUFile ssu_aln 20

!cd $workDir; ./ssu_aln.ssu-align.sh

!cd $workDir; ssu-mask --dna --afa ssu_aln

#!cd $workDir; ~/seq_tools/seq_tools fasta_unwrap ssu_aln/ssu_aln.bacteria.mask.afa  | head

### Using Sulfolobus as the root (acc. X90478).
# Skip

sso_acc = "X90478"
sso_fa = Entrez.efetch(db="nucleotide", id=sso_acc, rettype="fasta", retmode="text").readlines()

out = open(os.path.join(workDir, 'X90478.fasta'), 'w')
sso_fa_namestrip = sso_fa[1:]
sso_fa_namestrip.insert(0,">%s\n"%sso_acc)
out.writelines(sso_fa_namestrip)
out.close()

 !cd $workDir; head -n 4 X90478.fasta

### Align the outgroup to the same cm (use the same mask that was established earlier).

!cd $workDir; ssu-align -f -n bacteria --dna --rfonly X90478.fasta sso_aln

!cd $workDir; ssu-mask -s ssu_aln/ssu_aln.bacteria.mask --dna --afa sso_aln/

#!cd $workDir; /home/bryan/seq_tools/seq_tools fasta_unwrap sso_aln/sso_aln.bacteria.mask.afa | head

!cd $workDir; cat sso_aln/sso_aln.bacteria.mask.afa ssu_aln/ssu_aln.bacteria.mask.afa > aln_for_tree_wSulfo.fasta
!cd $workDir; cp ssu_aln/ssu_aln.bacteria.mask.afa  aln_for_tree.fasta

#!cd $workDir; /home/bryan/seq_tools/seq_tools fasta_unwrap aln_for_tree.fasta | head

!cd $workDir; head aln_for_tree_wSulfo.fasta

## Inferring and rooting the tree

aln = LoadSeqs(os.path.join(workDir, 'aln_for_tree.fasta'), moltype=DNA)
t_unroot = build_tree_from_alignment(aln, moltype=DNA)

#t_rooted = t_unroot.rootedWithTip('X90478')

t_unroot.writeToFile(os.path.join(workDir, 'Master_unrooted.tree'))

# Tree with sulfo

aln = LoadSeqs(os.path.join(workDir, 'aln_for_tree_wSulfo.fasta'), moltype=DNA)
t_unroot = build_tree_from_alignment(aln, moltype=DNA)

t_rooted = t_unroot.rootedWithTip('X90478')

t_rooted.writeToFile(os.path.join(workDir, 'Master_wSulfo.tree'))
