==> ms_notebooks/4-FinalizePhyloseqObject.ipynb <==
# Finalize merged dataset 
* Incorporate phylogenetic tree into phyloseq object
* Save master and split versions

library(dada2)
library(tidyr)
library(dplyr)
library(phyloseq)
library(ggplot2)
library(ape)

ps.thresh = readRDS('~/Hyphosphere/data/3Exp/phyloseq/3Exp_phyloseq_thresh.rds')
ps.thresh

# Remove duplicate samples from experiment 3

samples = sample_names(ps.thresh)
# These samples have higher contaminants than the replicates on plate 1 and so
# rather than merging the samples they are removed here
Duplicates = c('155_1.3RH', '156_2.3RH','157_3.3RH','158_1.4RH', '159_2.4RH')
samplestokeep = samples[!(samples %in% Duplicates)]
ps.thresh = prune_samples(samplestokeep, ps.thresh)
sample_names(ps.thresh)
# allTaxa = taxa_names(ps)
# allTaxa <- allTaxa[!(allTaxa %in% FC)]
# ps.sub = prune_taxa(allTaxa, ps)
# ps.thresh %>% subset_samples(Experiment == 3) %>% subset_samples(!sample_names(ps.thresh) %in% ) %>% sample_names()

thresh.tree = read_tree(treefile = '~/Hyphosphere/data/3Exp/Fasttree/Master_wSulfo.tree')
#is.rooted(thresh.tree)
thresh.tree = root(thresh.tree, outgroup = 'X90478', r = TRUE)
is.rooted(thresh.tree)
thresh.tree

# root.by.outgroup <- function(tree.unrooted) {
#   longest.edge <- which.max(tree.unrooted$edge.length)
#   long.nodes <- tree.unrooted$edge[longest.edge,]     #this should usually include one tip
#   new.outgroup <- long.nodes[long.nodes < Ntip(tree.unrooted)]
#   tree.rooted <- root(tree.unrooted, outgroup=new.outgroup, resolve.root=T)
#   tree.rooted
# }

# require(ape)
# thresh.tree = root.by.outgroup(thresh.tree)




# join bacterial tree with physeq object

ps = merge_phyloseq(ps.thresh, thresh.tree)
#force polytomies to dichotomies
phy_tree(ps) = multi2di(phy_tree(ps), random = FALSE)
ps

#remove mitochondria
ps = subset_taxa(ps, Family != "Mitochondria" | is.na(Family))
ps = subset_taxa(ps, Order != "Chloroplast" | is.na(Order))
ps

saveRDS(ps, '~/Hyphosphere/data/3Exp/phyloseq/final/3Exp_phyloseq_thresh_wtree.rds')


## save with contaminants removed see notebook 6

FC = c("ASV1", 
"ASV54" ,
"ASV78", 
"ASV136", 
"ASV3", 
"ASV12",
"ASV2516",
"ASV681", 
"ASV447",
"ASV49", 
"ASV12695", 
"ASV2395", 
"ASV284", 
"ASV1706", 
"ASV1900", 
"ASV26", 
"ASV14332", 
"ASV2962",
"ASV9101",
"ASV2180", 
"ASV8061", 
"ASV4437", 
"ASV6987", 
"ASV2714",
"ASV2160", 
"ASV2117", 
"ASV2704", 
"ASV1002", 
"ASV4083",
"ASV7871", 
"ASV1900", 
"ASV171", 
"ASV2516", 
"ASV8536",
"ASV285", 
"ASV3225", 
"ASV5265", 
"ASV2962", 
"ASV14282")

allTaxa = taxa_names(ps)
allTaxa <- allTaxa[!(allTaxa %in% FC)]
ps.sub = prune_taxa(allTaxa, ps)
# new phyloseq object with just the taxa you kept.
ps.sub


saveRDS(ps.sub, '~/Hyphosphere/data/3Exp/phyloseq/final/3Exp_phyloseq_thresh_wtree_woContam.rds')

### merge duplicates and save experimental samples

ps.sub.m = ps.sub %>% merge_samples(., "Sample")

### re-map factors

sample_data(ps.sub.m)$Sample <- sample_names(ps.sub.m)
sample_data(ps.sub.m)$Fungus <- levels(sample_data(ps.sub)$Fungus)[get_variable(ps.sub.m, "Fungus")]
sample_data(ps.sub.m)$Plant <- levels(sample_data(ps.sub)$Plant)[get_variable(ps.sub.m, "Plant")]
sample_data(ps.sub.m)$Soil <- levels(sample_data(ps.sub)$Soil)[get_variable(ps.sub.m, "Soil")]
sample_data(ps.sub.m)$SampleType <- levels(sample_data(ps.sub)$SampleType)[get_variable(ps.sub.m, "SampleType")]
sample_data(ps.sub.m)$Treatment <- levels(sample_data(ps.sub)$Treatment)[get_variable(ps.sub.m, "Treatment")]

sample_data(ps.sub.m)[sample_data(ps.sub.m)$Experiment == 3,] %>% head()

sample_data(ps.sub.m)$SampleType %>% unique()

outfile = ps.sub.m %>% subset_samples(., 
            !SampleType %in% c("MK", "EB", "PB", "T0")) %>%
            subset_samples(!Sample == "Unknown-RS")
outfile

saveRDS(outfile, '~/Hyphosphere/data/3Exp/phyloseq/final/3Exp_phyloseq_thresh_wtree_experimental_woContam.rds')

# now larger object without tree

allTaxa = taxa_names(ps.thresh)
allTaxa <- allTaxa[!(allTaxa %in% FC)]
ps.sub = prune_taxa(allTaxa, ps.thresh)
# new phyloseq object with just the taxa you kept.
ps.sub

#remove mitochondria
ps.sub = subset_taxa(ps.sub, Family != "Mitochondria" | is.na(Family))
ps.sub = subset_taxa(ps.sub, Order != "Chloroplast" | is.na(Order))
ps.sub

### merge duplicates and save experimental samples

ps.sub.m = ps.sub %>% merge_samples(., "Sample")

### re-map factors

sample_data(ps.sub.m)$Sample <- sample_names(ps.sub.m)
sample_data(ps.sub.m)$Fungus <- levels(sample_data(ps.sub)$Fungus)[get_variable(ps.sub.m, "Fungus")]
sample_data(ps.sub.m)$Plant <- levels(sample_data(ps.sub)$Plant)[get_variable(ps.sub.m, "Plant")]
sample_data(ps.sub.m)$Soil <- levels(sample_data(ps.sub)$Soil)[get_variable(ps.sub.m, "Soil")]
sample_data(ps.sub.m)$SampleType <- levels(sample_data(ps.sub)$SampleType)[get_variable(ps.sub.m, "SampleType")]
sample_data(ps.sub.m)$Treatment <- levels(sample_data(ps.sub)$Treatment)[get_variable(ps.sub.m, "Treatment")]

sample_data(ps.sub.m)[sample_data(ps.sub.m)$Experiment == 3,] %>% head()

sample_data(ps.sub.m)$SampleType %>% unique()

outfile = ps.sub.m %>% subset_samples(., 
            !SampleType %in% c("MK", "EB", "PB", "T0")) %>%
            subset_samples(!Sample == "Unknown-RS")
outfile

saveRDS(outfile, '~/Hyphosphere/data/3Exp/phyloseq/final/3Exp_phyloseq_thresh_woTree_experimental_woContam.rds')

sessionInfo()
