==> ms_notebooks/5-Ordination_and_hypothesis_testing.ipynb <==
# Generate ordinations and permanovas for hyphal communities manuscript

figDir = '~/Hyphosphere/data/3Exp/figures/final'

# phyloseq - full object with tree (does not include archaea); needs MREs removed
physeq_file = '~/Hyphosphere/data/3Exp/phyloseq/final/3Exp_phyloseq_thresh_wtree_experimental_woContam.rds'

library(RColorBrewer)
library(vegan)
library(ggpubr)
library(tidyr)
library(dplyr)
library(reshape2)
library(phyloseq)
library(gridExtra)
library(ggplot2)
library(foreach)
library(doParallel)


theme_pub = function (base_size = 8, base_family = "Helvetica") 
{
    theme_grey(base_size = base_size, base_family = base_family) %+replace% 
        theme(
            axis.line.x = element_line(size = .5),
            axis.line.y = element_line(size = .5),
            axis.text = element_text(size = base_size),
            panel.background = element_rect(fill = "white", colour = NA), 
            panel.border = element_rect(fill = NA, color = "black", size = .5), 
            panel.grid.major = element_blank(), 
            panel.grid.minor = element_blank(), 
            strip.background = element_rect(fill = "grey85", 
                colour = "grey20"), 
            legend.key = element_rect(fill = "white", 
                colour = NA), complete = TRUE,
            legend.text = element_text(size = base_size),
            legend.title = element_text(face = "bold"))
    
    }

ps = readRDS(physeq_file)
ps = subset_taxa(ps, !is.na("Phylum"))


tax_table(ps) %>% head()
ps = subset_taxa(ps, Kingdom == "Bacteria" &  !is.na(Phylum))
ps = subset_samples(ps, !SampleType %in% c("EB", "PB", "MK")) #remove non-experimental controls

# re-normalize after subsetting taxa
ps.snorm = transform_sample_counts(ps, function(x) x/sum(x))

otu_table(ps.snorm)[1:4, 1:10]
tax_table(ps.snorm) %>% tail()

ps.snorm
                                  


# Experiment 1 & 2 primary analysis

ps.s = ps.snorm %>% subset_samples(Experiment %in% c(1, 2) & 
                                   SampleType %in% c("CS", "CH", "BS") & Fungus == "Gv" & Soil != "Sand" &
                                  Treatment == 'HN')

registerDoParallel(makeCluster(20))

wunif.dist = phyloseq::distance(ps.s, 
                      method = "unifrac", 
                      weighted = TRUE,
                      fast = TRUE, 
                      parallel = TRUE, 
                      normalized = FALSE)


set.seed(101)

df = as(sample_data(ps.s), "data.frame")
df$Rep = as.factor(df$Rep)
df$TimePoint = as.factor(df$TimePoint)
d = wunif.dist

adonis.w = adonis(formula = d ~Soil + SampleType + SampleType:Soil + Experiment + Experiment:SampleType + Soil:Experiment, df, permutations = 999)
print("weighted-UniFrac")
adonis.w

ordu = ordinate(ps.s, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.s, ordu, justDF = TRUE)
plot_ordination(ps.s, ordu)

#setwd(figDir)
cs = brewer.pal(3,"Dark2")
p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(shape = as.factor(Soil), fill = SampleType, color = as.factor(Experiment)),  size = 3, stroke = 1.5) +
theme_pub() +
scale_shape_manual(values=c(22,21,23), guide = guide_legend(title = "Soil", ncol = 1))+
scale_color_manual(values=c('Orange','Black'))+
scale_fill_manual(values = cs, labels = c("Bulk soil", "Core hyphae", "Core slurry"))+
labs(x = "PCoA 1 [37.0%]", y = "PCoA 2 [12.9%]") +
ggtitle("PCoA of Weighted UniFrac Distances")+
#facet_wrap(~Experiment, ncol = 2)

guides(fill=guide_legend(override.aes=list(colour= cs), ncol = 3)) + 
theme(legend.position="bottom") 

#color = "grey20",

p1

pdf(file.path(figDir,"Exp1-2_wUnif_PCOA_BSCHCS.pdf"), useDingbats = FALSE, width = 4.5, height = 4.5)
p1
dev.off()

# Compare bulk soil vs bulk hyphosphere
* not presented in ms

ps.b = subset_samples(ps.snorm, SampleType %in% c("BS", "BH") & Experiment == 2 & Sample != "GVN7HP4-BH")

b.wunif.dist = phyloseq::distance(ps.b, 
                      method = "unifrac", 
                      weighted = TRUE,
                      fast = TRUE, 
                      parallel = TRUE, 
                                  
                      normalized = FALSE)

ordu = ordinate(ps.b, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.b, ordu, justDF = TRUE)
plot_ordination(ps.b, ordu)


cs = brewer.pal(7,"Dark2")
p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(shape = as.factor(Soil), fill = SampleType),  size = 2) +
theme_pub() +
scale_shape_manual(values=c(21,22,23), guide = guide_legend(title = "Soil", ncol = 1))+
scale_fill_manual(values = cs[c(1,3)], labels = c("Bulk hyphosphere", "Bulk soil"))+
labs(x = "PCoA 1 [41.6%]", y = "PCoA 2 [13.1%]") +
ggtitle("PCoA of Weighted UniFrac Distances")+

guides(fill=guide_legend(override.aes=list(colour= cs[c(1,3)]), ncol = 1)) + 
theme(legend.position="bottom") +
facet_wrap(~Fungus, ncol = 2)



p1

#pdf(file.path(figDir,"wUnif_PCOA_BSBH.pdf"), useDingbats = FALSE, width = 4.5, height = 3)
#p1
#dev.off()

set.seed(101)

df = as(sample_data(ps.b), "data.frame")
df$Rep = as.factor(df$Rep)
df$TimePoint = as.factor(df$TimePoint)
d = b.wunif.dist

b.adonis.w = adonis(formula = d ~ Rep + Soil + SampleType + Soil:SampleType + Fungus + Fungus:SampleType, df, permutations = 999)
print("weighted-UniFrac")
b.adonis.w



# Additional controls - Fig. S4

ps.s = ps.snorm %>% subset_samples(Experiment == 2 & Fungus == "Gv" & SampleType != "RT")

ordu = ordinate(ps.s, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.s, ordu, justDF = TRUE)
plot_ordination(ps.s, ordu)

brewer.pal(7,"Dark2")

cs = c('#66a61e', #BH
       '#e7298a', #BP
       '#1b9e77', #BS
       
        '#d95f02', #CH, 
       '#7570b3', #CS,
       "#e6ab02", #RH
       "#a6761d") #RS



p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(shape = Soil, fill = SampleType,  size = 2, stroke = 1.2)) +
theme_pub() +
scale_shape_manual(values=c(21,22,23), guide = guide_legend(title = "Soil", ncol = 2))+
scale_fill_manual(values = cs)+
#scale_color_manual(values=c('Light Grey','Black'))+
labs(x = "PCoA 1 [44.2%]", y = "PCoA 2 [12.2%]") +
#ggtitle("PCoA of Weighted UniFrac Distances")+

guides(fill=guide_legend(override.aes=list(colour= cs), ncol = 4)) + 
theme(legend.position="bottom") 

#color = "grey20",

p1

pdf(file = file.path(figDir,'Exp1_2_FullOrdination.pdf'), width = 7.086, height = 5, useDingbats = FALSE)
p1
dev.off()


# Test fungal species 
* Fig. 3a and Table 2

ps.s = subset_samples(ps.snorm, SampleType %in% c("BS","CS","CH") & Soil == "Dryden" & Experiment == 2)

sample_data(ps.s) %>% head()

ordu = ordinate(ps.s, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.s, ordu, justDF = TRUE)
plot_ordination(ps.s, ordu)

cs = brewer.pal(3,"Dark2")
p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(shape = as.factor(Fungus), fill = SampleType),  size = 2, stroke = 1.5) +
theme_pub() +
scale_shape_manual(values=c(21,22), guide = guide_legend(title = "Treatment", ncol = 1))+
scale_fill_manual(values = cs, labels = c("Bulk soil", "Core hyphae", "Core slurry"))+
labs(x = "PCoA 1 [70%]", y = "PCoA 2 [8.9%]") +

guides(fill=guide_legend(override.aes=list(colour= cs, ncol = 1))) + 
theme(legend.position="bottom")

#color = "grey20",

p1
pf = p1
pdf(file.path(figDir,"wUnif_PCOA_CHCSxFungus.pdf"), useDingbats = FALSE, width = 3.14, height = 3)
p1
dev.off()
rm(p1)

s.wunif.dist = phyloseq::distance(ps.s, 
                      method = "unifrac", 
                      weighted = TRUE,
                      fast = TRUE, 
                      parallel = TRUE, 
                      normalized = FALSE)

### set.seed(101)

df = as(sample_data(ps.s), "data.frame")
df$Rep = as.factor(df$Rep)
df$TimePoint = as.factor(df$TimePoint)
d = s.wunif.dist

s.adonis.w = adonis(formula = d ~ as.factor(Rep) + SampleType + Fungus + Fungus:SampleType, df, permutations = 999)
print("weighted-UniFrac")
s.adonis.w

tabout = s.adonis.w$aov.tab
tabout[,2:5] = round(tabout[,2:5], 2)
tabout
fileout = file.path(figDir, "ADONIS_Exp2_FungalSpecies.csv")
write.table(tabout, file = fileout, sep = ',', row.names = TRUE)

# Test nutrient treatment
* Fig. S8 

ps.b = subset_samples(ps.snorm, SampleType %in% c("CH") & Soil != "Pendelton" & Experiment == 2)
ordu = ordinate(ps.b, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.b, ordu, justDF = TRUE)
plot_ordination(ps.snorm, ordu)

cs = brewer.pal(3,"Dark2")[2]
p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(fill = Treatment, shape = Soil),  size = 2, color = cs) +
theme_pub() +
scale_shape_manual(values=c(21,22,23,24), guide = guide_legend(title = "Soil", ncol = 1))+
scale_fill_manual(values = c(cs, "White"))+
labs(x = "PCoA 1 [36.7%]", y = "PCoA 2 [22.1%]") +
ggtitle("PCoA of Weighted UniFrac Distances")+

guides(fill=guide_legend(override.aes=list(color= c("Grey", "White")), ncol = 1)) + 
theme(legend.position="bottom")
#facet_wrap(Fungus~Soil, ncol = 2)

#color = "grey20",

p1



b.wunif.dist = phyloseq::distance(ps.b, 
                      method = "unifrac", 
                      weighted = TRUE,
                      fast = TRUE, 
                      parallel = TRUE, 
                      normalized = FALSE)

df = as(sample_data(ps.b), "data.frame")
#df$Rep = as.factor(df$Rep)
#df$TimePoint = as.factor(df$TimePoint)
d = b.wunif.dist

b.adonis.w = adonis(formula = d ~ as.factor(Rep) + Soil + Fungus + Treatment + Treatment:Fungus + Treatment:Soil, df, permutations = 999)
print("weighted-UniFrac")
b.adonis.w
Atab = b.adonis.w$aov %>% as.data.frame()
Atab[,2:5] = round(Atab[,2:5], digits=2) 
Atab = Atab %>% select(-SumsOfSqs)
Atab


pm1 = ggtexttable(Atab, theme = ttheme("blank", base_size = 8))
ggarrange(p1, pm1, ncol = 2, widths = c(1.5, 1))

pdf(file.path(figDir,"wUnif_PCOA_CHxTreatment.pdf"), useDingbats = FALSE, width = 7.31, height = 4)
ggarrange(p1, pm1, ncol = 2)
dev.off()

# Test temporal trends
* Fig. 3b and Table 3

ps.t = subset_samples(ps.snorm, SampleType %in% c("CH", "CS", "BS") & Experiment == 3)
ordu = ordinate(ps.t, "PCoA", "unifrac", weighted=TRUE)
p = plot_ordination(ps.t, ordu, justDF = TRUE)
plot_ordination(ps.snorm, ordu)

cs = brewer.pal(5,"YlGn")
p1 = ggplot(p, aes(x = Axis.1, y = Axis.2)) +
geom_point(aes(shape = as.factor(SampleType), fill = as.factor(TimePoint)),  size = 2) +
theme_pub() +
scale_shape_manual(values=c(21,22,23), guide = guide_legend(title = "Sample Type", ncol = 1))+
scale_fill_manual(values = cs)+
labs(x = "PCoA 1 [49.4%]", y = "PCoA 2 [19.1%]") +

guides(fill=guide_legend(override.aes=list(colour= cs), ncol = 2)) + 
theme(legend.position="bottom") 
p1
pt = p1

pdf(file.path(figDir,"wUnif_PCOA_time.pdf"), useDingbats = FALSE, width = 3.14, height = 3)
p1
dev.off()
rm(p1)

t.wunif.dist = phyloseq::distance(ps.t, 
                      method = "unifrac", 
                      weighted = TRUE,
                      fast = TRUE, 
                      parallel = TRUE, 
                      normalized = FALSE)

df = as(sample_data(ps.t), "data.frame")
#df$Rep = as.factor(df$Rep)
#df$TimePoint = as.factor(df$TimePoint)
d = t.wunif.dist

t.adonis.w = adonis(formula = d ~ as.factor(Rep) + SampleType + as.factor(DAI) + as.factor(DAI):SampleType , df, permutations = 999)
print("weighted-UniFrac")
t.adonis.w

tabout = t.adonis.w$aov.tab
tabout[,2:5] = round(tabout[,2:5], 2)
tabout



fileout = file.path(figDir, "ADONIS_Exp3_CHCSBS.csv")
write.table(tabout, file = fileout, sep = ',', row.names = TRUE)

# Multipanel for publication


require(cowplot)

pdf(file.path(figDir,"MultPanel2_a-b.pdf"), useDingbats = FALSE, width = 7.086, height = 3.3)
plot_grid(pf, pt, ncol = 2, align = "h")
dev.off()
